textUtils: rename module
