+++ src/text_utils.js
